refactor: Game, Core class 추가 Game - start 메서드 : 게임 시작 - startGame 메서드 : 게임 시작 문구 출력 - getUserInput 메서드 : 사용자의 input을 입력 받기 - isRestart 메서드 : 사용자의 재시작 여부 input을 입력 받기 - isEnd 메서드 : 게임의 종료 조건 검사
Core
- strike : strike 개수를 리턴
- ball : ball 개수를 리턴
- printResult : 사용자의 input에 대한 결과 출력
